Implement initial support for stencil buffering.
glClipPlane (vertex based user clipping) isn't supported in GLES
or in combination with GL2's vertex shaders. As such, we need
to replace it with a more modern alternative.

A wrapable function was added glBufferStencil which provides
plugins with a GLVertexBuffer, and readonly GLMatrix and
CompOutput. This allows plugins to draw in areas where stenciling
will pass using the passed GLVertexBuffer.