Fix regression; glDrawTexture was being called MUCH more than it does with trunk. (LP: #1029806)
This was because Linaro removed GLWindow::Geometry, and the decision to call
glDrawTexture used to be conditional on GLWindow::Geometry::vCount. So they
just removed the if's and made it unconditional, hence slower.

Now I've reintroduced GLWindow::Geometry::vCount in the form of
GLVertexBuffer::countVertices, and glDrawTexture is only called conditionally
again.

This will also make it possible/easy to reintroduce some deformation features
that are presently commented out (by Linaro). We will just have to convert:
    GLWindow::Geometry::vCount
    GLWindow::Geometry::vertices
to:
    GLVertexBuffer::countVertices()
    GLVertexBuffer::getVertices()