The grid resize hint overlay doesn't disappear when the window is moved along the edge from a region that has a resize action set to a region without one (LP: #839602).
When the edge changes, grid calculates the target rectangle for the
transformation. The previous overlay will be destroyed only if the
new target rectangle is different from the previous target rectangle.
But when the target is GridUnknown (the action for the region has set
to be None), initiateCommon bails out right away leaving the target
rectangle unchanged. And thus grid thinks nothing has changed and
the old overlay is left intact.

The fix resets the target rectangle to be at (0, 0) with height and
width of 0. As the previous rectangle is now different from the new
one, the old overlay will fade out. The same technique is used to
destroy the overlay when the window is moved away from the target region.