Expose ccsCopyInfo and separate ccsCopyValueInto from ccsCopyValue. The latter returns a newly allocated copy of the value, the former deep-
There was a bug in ccsCopyInfo which caused the copy-from info to be
overwritten with the copy-to info. That was fixed.

There are no tests at present. The testing framework for this will be
quite similar to what we need to test settings, and will be adapted for
that purpose in testing this code.