Added some walking skeleton code for the new pixmap release system so that we can get both unit and integration tests in.
PixmapReleasePool - Textures that no longer need a pixmap can notify this object
through PixmapDestroyQueue that the pixmap can be destroyed. The decorator can
communicate a pixmap is unused through UnusedPixmapQueue. If the former happens
first, then the pixmap is stored here as "unusued" pending a call through
PixmapDestroyQueue to get rid of it. Otherwise, it will be destroyed immediately.

decor::protocol::Communicator - Receive XClient messages and dispatch to:

PendingHandler - Receive messages and notify relevant decoratiosn that new
pixmaps are ready.

UnusedHandler - Receive messages and notify the UnusuedPixmapQueue that a
pixmap is ready to be freed.